Don't include "resolve" in @babel/standalone
