+++ packages/babel-plugin-transform-runtime/package.json
@@ -11,6 +11,9 @@
   "keywords": [
     "babel-plugin"
   ],
+  "browser": {
+    "./lib/get-runtime-path/index.js": "./lib/get-runtime-path/browser.js"
+  },
   "dependencies": {
     "@babel/helper-module-imports": "^7.8.3",
     "@babel/helper-plugin-utils": "^7.8.3",
