+++ packages/babel-plugin-transform-runtime/src/get-runtime-path/browser.js
@@ -1,0 +1,7 @@
+export default function(moduleName, dirname, absoluteRuntime) {
+  if (absoluteRuntime === false) return moduleName;
+
+  throw new Error(
+    "The 'absoluteRuntime' option is not supported when using @babel/standalone.",
+  );
+}
